refactor Notifications.kt
