+++ src/main/kotlin/io/github/eone666/telegramnotifier/TelegramNotifier.kt
@@ -2,7 +2,7 @@ package io.github.eone666.telegramnotifier
 
 import io.github.eone666.telegramnotifier.listeners.PlayerJoin
 import io.github.eone666.telegramnotifier.listeners.PlayerQuit
-import io.github.eone666.telegramnotifier.features.notifications.Notifications
+import io.github.eone666.telegramnotifier.features.Notifications
 import io.github.eone666.telegramnotifier.features.PlayersList
 import io.github.eone666.telegramnotifier.listeners.PlayerAdvancement
 import io.github.eone666.telegramnotifier.utils.Config
