--- src/main/kotlin/io/github/eone666/telegramnotifier/features/notifications/NotificationTypes.kt
@@ -1,6 +1,0 @@
-package io.github.eone666.telegramnotifier.features.notifications
-
-enum class NotificationTypes {
-    JOIN,
-    QUIT
-}
