+++ src/main/kotlin/io/github/eone666/telegramnotifier/utils/telegram/ParseMode.kt
@@ -1,8 +1,8 @@
 package io.github.eone666.telegramnotifier.utils.telegram
 
 enum class ParseMode(private val text: String) {
-    MARKDOWN("markdown");
-//    HTML("html");
+    MARKDOWN("Markdown"),
+    HTML("HTML");
 
     override fun toString(): String {
         return text
